+++ packages/hds/src/systems/fonts/PretendardVariable.woff2
